+++ src/main/java/com/KHH/owner_myPage/AddRestaurantC.java
@@ -1,7 +1,6 @@
 package com.KHH.owner_myPage;
 
 import java.io.IOException;
-import java.rmi.ServerException;
 import javax.servlet.ServletException;
 import javax.servlet.annotation.WebServlet;
 import javax.servlet.http.HttpServlet;
@@ -12,8 +11,9 @@ import javax.servlet.http.HttpServletResponse;
 public class AddRestaurantC extends HttpServlet {
 
     protected void doGet(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
-        Owner_DAO.InsertRestaurant(request);
 
+        Restaurant_DAO.InsertRestaurant(request);
+        request.setAttribute("content","");
     }
 
     protected void doPost(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
