+++ src/main/java/com/KHH/owner_myPage/ShowAllRestaurantC.java
@@ -1,7 +1,6 @@
 package com.KHH.owner_myPage;
 
 import java.io.IOException;
-import java.rmi.ServerException;
 import javax.servlet.ServletException;
 import javax.servlet.annotation.WebServlet;
 import javax.servlet.http.HttpServlet;
@@ -12,8 +11,8 @@ import javax.servlet.http.HttpServletResponse;
 public class ShowAllRestaurantC extends HttpServlet {
 
     protected void doGet(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
-        Owner_DAO.ShowAllList(request);
-
+        Restaurant_DAO.ShowAllList(request);
+        request.getRequestDispatcher("jsp/owner_myPage/ListMyRestaurant.jsp").forward(request, response);
     }
 
     protected void doPost(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
